Update software versions used in CI
The ubuntu-20.04 runner is due to be deprecated on 2025-02-01 and unsupported by 2025-04-01.

The previous version of the metadata validator wouldn't run on Ubuntu 24.04 due to a dependency version mismatch, so has been updated.

The actions/checkout and actions/setup-node actions have also been updated to their latest major versions.

The version of Node.js used has also been updated to the latest LTS version.